Improve resilience of SyncWorker
It's the bread and butter of this whole thing, so it needs to be pretty
solid. To that end, I have brought in Mixlib::Shellout which hopefully
gives us a cross-platform bazooka's worth of options which will simplify
managing subprocesses.

Adding timeouts should help with any other Sketch bugs that hang the
process indefinitely.

Also, instead of using `retry`, I altered the logic to enqueue the same
sync job at the end of the queue so it would not block jobs in the
meanwhile.