WIP: Rewrite JS in React
This is pretty much a direct port of the janky jQuery DOM manipulation
into a poorly-structured, <div>-heavy React component topology.

It uses react-waypoint to replace the jQuery lazy loading, and
react-router to handle the inter-page navigation, which should hopefully
give me a better place to go from when I want to add new pages to the
site.

TODO:
- Cleanup the parts where React eats my whitespace
- Rebuild the tags page (add that logic back in)
- Use Bower to manage vendored JS
  - Somewhat blocked by
    https://github.com/brigade/react-waypoint/issues/13
  - Unless we:
- Use some CommonJS builder (Webpack, Browserify, etc.) to compile JSX /
  minify JS / etc. (I explicitly want to avoid concatenating all JS into
  a single file for as long as possible, though. Mosaic is meant to be
  served over HTTP/2.)