Improve searching performance
This is probably mostly a placebo, but there are two changes here:

First, this debounces the search bar to one query every 100 ms. This
should help reduce the amount of work browsers have to do without
affecting the user experience too much. (Fixes #12)

Second, this throws away state changes to the Search page if they aren't
relevant to the query the user entered. I haven't fully reasoned through
this in my head, so it might contain race conditions, but the idea is
that as the user types quickly, or if the user has hit a key and
backspace quickly, then there is no need to process the intermediary
search results because they are not the response for what exists in the
search box right now.