Add "/" keyboard shortcut to focus search (#9)
This is pretty useless on its own because the search bar autofocuses
anyhow. But, baby steps.

Ideally, you'd be able to traverse through search results with a
shortcut sequence like:

  /campaign<tab>jjjlllll

Where j/k move between files and h/l move between slices in that file.
Or something like that. I'm going to need to think a bit more about how
I can restructure the react so that I can actually make the search
results accessible when the search bar is blurred.