Add server URL to readme
Once you start up the app for the first time, it starts downloading
everything from Dropbox and parsing it out, filling up the terminal. It
wasn't obvious to me how to access my local instance once I had it
running, so I figured it would be good to add this information to the
readme.