Make result images responsive
On smaller screens, these images would be clipped horizontally, which is
less than ideal. Setting a max-width of 100% fixes this, improving the
usability of the app.

At the same time, I am removing the data-original attribute from the
image because it does not appear to be used anymore. I believe that this
may have been a holdover from Mosaic's pre-React days.