Restructure app to be component-centric
Now that we are using style-loader and css-loader, we can make each
component a nice little bundle of JSX and styles. This commit simply
restructures the js directory to be called components and gives each
component its own directory which can have JSX and CSS. In my next
commit I will distribute the styles to each component directory.