Move spacer gap styling to <Header>
This keeps the styles for this in the same place, which should make
things a little easier to understand. The drawback is that this
component could be rendered anywhere but this height only makes sense if
it is rendered at the top of the page. However, in practice that is the
only place it is rendered and it seems unlikely that will change, so
this seems fine for now.